refactor[domain]: extrai campos de endereço para classe Endereco e atualiza ProfissionalSaude
